+++ packages/components/organisms/f-checkout/test-utils/component-objects/f-checkout.component.js
@@ -109,15 +109,13 @@ module.exports = class Checkout extends Page {
      * @param {String} checkout.isValid The checkout validation
      */
 
-    open (checkout) {
-        const serviceType = checkout.isValid ? `&knob-Service%20Type=${checkout.type}` : '&knob-Service%20Type=Invalid%20URL';
-        const isLoggedIn = `&knob-Is%20User%20Logged%20In=${checkout.isAuthenticated}`;
-        const hasCheckoutErrors = `&knob-Checkout%20Errors=${checkout.checkoutErrors}`;
-        const hasPlaceOrderErrors = `&knob-Place%20Order%20Errors=${checkout.placeOrderErrors}`;
-        const hasAsapAvailable = `&knob-Is%20ASAP%20available=${checkout.isAsapAvailable}`;
+    open (url) {
+        super.open(url);
+    }
 
-        const url = `checkout-component${serviceType}${isLoggedIn}${hasCheckoutErrors}${hasPlaceOrderErrors}${hasAsapAvailable}`;
-        super.openComponent('organism', url);
+    withQuery (name, value) {
+        super.withQuery(name, value);
+        return this;
     }
 
     waitForComponent () {
@@ -166,7 +164,7 @@ module.exports = class Checkout extends Page {
     * @param {String} userInfo.email The user's e-mail address
     * @param {String} userInfo.password The user's password
     */
-    submitForm () {
+    submitForm (userInfo) {
         this.fields.firstName.input.setValue(userInfo.firstName);
         this.fields.lastName.input.setValue(userInfo.lastName);
         this.fields.email.input.setValue(userInfo.email);