+++ packages/components/organisms/f-footer/test-utils/component-objects/f-footer.component.js
@@ -64,46 +64,8 @@ module.exports = class Footer extends Page {
         this.countryValue = this.countries.filter(element => element.getAttribute('data-test-id').includes(country))[0];
     }
 
-    /**
-     * @description
-     * Sets the data for the checkout component.
-     *
-     * @param {Object} footer
-     * @param {String} footer.locale The checkout type
-     * @param {String} footer.countrySelector The checkout authentication
-     * @param {String} footer.courierLinks The checkout authentication
-     */
-
-    open (footer) {
-        const countryFormatted = footer.locale.toUpperCase();
-        const showCountrySelector = `&knob-Show%20country%20selector=${footer.countrySelector}`;
-        const showCourierLinks = `&knob-Show%20courier%20links=${footer.courierLinks}`;
-        let formattedLocale = '';
-        switch (countryFormatted) {
-            case 'GB':
-            case 'AU':
-            case 'NZ':
-            case 'IE':
-                formattedLocale = `en-${countryFormatted}`;
-                break;
-            case 'DK':
-                formattedLocale = `da-${countryFormatted}`;
-                break;
-            case 'ES':
-                formattedLocale = `es-${countryFormatted}`;
-                break;
-            case 'IT':
-                formattedLocale = `it-${countryFormatted}`;
-                break;
-            case 'NO':
-                formattedLocale = `nb-${countryFormatted}`;
-                break;
-            default:
-                throw new Error(`locale ${countryFormatted} is not supported`);
-        }
-
-        const url = `footer-component&knob-Locale=${formattedLocale}${showCountrySelector}${showCourierLinks}`;
-        super.openComponent('organism', url);
+    open (url) {
+        super.open(url);
     }
 
     waitForComponent () {
@@ -153,7 +115,6 @@ module.exports = class Footer extends Page {
     areCourierLinksDisplayed () {
         return this.courierLinks.isDisplayed();
     }
-
     isCountrySelectorDisplayed () {
         return this.countrySelectorButton.isDisplayed();
     }