+++ packages/services/f-wdio-utils/src/page.object.js
@@ -22,4 +22,3 @@ class Page {
 }
 
 module.exports = Page;
-