+++ CHANGELOG.md
@@ -3,14 +3,21 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-v3.36.0
+v3.37.0
 ------------------------------
-*May 27, 2021*
+*May 28, 2021*
 
 ## Added
 - CircleCI config for running `build`, `lint` & `unit testing` steps individually locally.
 - Documentation into storybook guides as to how to run these tests through CircleCI CLI locally and what each task is setup to do.
 
+
+v3.36.0
+------------------------------
+*May 27, 2021*
+
+## Added
+- `URL Builder` to all components for testing different knobs against storybook
 
 v3.35.0
 ------------------------------